big change - removed downloader and uploader && update many official plugins.
- new tg_downloader func for fast download
- modified pdftools, ziptools, vc dl and other plugins inbolving downloader nd uploader

+++ plugins/audiotools.py
@@ -11,14 +11,12 @@ from datetime import datetime as dt
 from pathlib import Path
 from shlex import quote
 
-from pyUltroid.fns.tools import set_attributes
-
+from pyUltroid.fns.tools import _stdr
 from . import (
     LOGS,
     ULTConfig,
     bash,
     check_filename,
-    downloader,
     eod,
     eor,
     genss,
@@ -27,11 +25,12 @@ from . import (
     humanbytes,
     mediainfo,
     osremove,
-    stdr,
+    set_attributes,
+    tg_downloader,
     time_formatter,
     ultroid_cmd,
-    uploader,
 )
+
 
 __doc__ = get_help("help_audiotools")
 
@@ -78,64 +77,49 @@ async def to_voice(e):
 
 @ultroid_cmd(pattern="atrim( (.*)|$)")
 async def trim_aud(e):
-    sec = e.pattern_match.group(1).strip()
+    sec = e.pattern_match.group(2)
     if not sec or "-" not in sec:
-        return await eod(e, get_string("audiotools_3"))
+        return await e.eor(get_string("audiotools_3"), time=5)
     a, b = sec.split("-")
     if int(a) >= int(b):
-        return await eod(e, get_string("audiotools_4"))
-    vido = await e.get_reply_message()
-    if vido and vido.media and mediainfo(vido.media).startswith(("video", "audio")):
-        if hasattr(vido.media, "document"):
-            vfile = vido.media.document
-            name = vido.file.name
-        else:
-            vfile = vido.media
-            name = ""
-        if not name:
-            name = dt.now().isoformat("_", "seconds") + ".mp4"
-        xxx = await e.eor(get_string("audiotools_5"))
-        c_time = time.time()
-        file = await downloader(
-            f"resources/downloads/{name}",
-            vfile,
-            xxx,
-            c_time,
-            f"Downloading {name}...",
-        )
+        return await e.eor(get_string("audiotools_4"), time=5)
 
-        o_size = Path(file.name).stat().st_size
-        d_time = time.time()
-        diff = time_formatter((d_time - c_time) * 1000)
-        file_name = (file.name).split("/")[-1]
-        out = file_name.replace(file_name.split(".")[-1], "_trimmed.aac")
-        if int(b) > int(await genss(file.name)):
-            osremove(file.name)
-            return await eod(xxx, get_string("audiotools_6"))
-        ss, dd = stdr(int(a)), stdr(int(b))
-        xxx = await xxx.edit(
-            f"Downloaded `{file.name}` of `{humanbytes(o_size)}` in `{diff}`.\n\nNow Trimming Audio from `{ss}` to `{dd}`..."
-        )
-        cmd = f'ffmpeg -i "{file.name}" -preset ultrafast -ss {ss} -to {dd} -vn -acodec copy "{out}" -y'
-        await bash(cmd)
-        osremove(file.name)
-        f_time = time.time()
-        mmmm = await uploader(out, out, f_time, xxx, f"Uploading {out}...")
-        attributes = await set_attributes(out)
+    reply = await e.get_reply_message()
+    if not (
+        reply and reply.media and mediainfo(reply.media).startswith(("video", "audio"))
+    ):
+        return await e.eor(get_string("audiotools_1"), time=5)
 
-        caption = get_string("audiotools_7").format(ss, dd)
-        await e.client.send_file(
-            e.chat_id,
-            mmmm,
-            thumb=ULTConfig.thumb,
-            caption=caption,
-            attributes=attributes,
-            force_document=False,
-            reply_to=e.reply_to_msg_id,
-        )
-        await xxx.delete()
-    else:
-        await e.eor(get_string("audiotools_1"), time=5)
+    xxx = await e.eor(get_string("audiotools_5"))
+    path, t_time = await tg_downloader(media=reply, event=xxx, show_progress=True)
+    o_size = Path(path).stat().st_size
+    diff = time_formatter(t_time * 1000)
+    out = check_filename(Path(path).with_suffix(".aac"))
+    if int(b) > int(await genss(path)):
+        osremove(path)
+        return await xxx.edit(get_string("audiotools_6"))
+
+    ss, dd = _stdr(int(a)), _stdr(int(b))
+    await xxx.edit(
+        f"Downloaded `{path}` of `{humanbytes(o_size)}` in `{diff}`.\n\nNow Trimming Audio from `{ss}` to `{dd}`..."
+    )
+    await asyncio.sleep(2)
+    cmd = f"ffmpeg -i {quote(path)} -preset ultrafast -ss {ss} -to {dd} -vn -acodec copy {quote(out)} -y"
+    await bash(cmd)
+    attributes = await set_attributes(out)
+    mmmm, _ = await e.client.fast_uploader(
+        out, show_progress=True, event=xxx, save_cache=False
+    )
+    caption = get_string("audiotools_7").format(ss, dd)
+    await reply.reply(
+        caption,
+        file=mmmm,
+        thumb=ULTConfig.thumb,
+        attributes=attributes,
+        force_document=False,
+    )
+    osremove(out, path)
+    await xxx.delete()
 
 
 @ultroid_cmd(pattern="extractaudio$")
@@ -143,34 +127,24 @@ async def ex_aud(e):
     reply = await e.get_reply_message()
     if not (reply and reply.media and mediainfo(reply.media).startswith("video")):
         return await e.eor(get_string("audiotools_8"))
-    name = reply.file.name or "video.mp4"
-    vfile = reply.media.document
+
     msg = await e.eor(get_string("com_1"))
-    c_time = time.time()
-    file = await downloader(
-        f"resources/downloads/{name}",
-        vfile,
-        msg,
-        c_time,
-        f"Downloading {name}...",
-    )
+    path, _ = await tg_downloader(media=reply, event=msg, show_progress=True)
+    out_file = Path(path).with_suffix(".aac")
+    cmd = f"ffmpeg -i {quote(path)} -vn -acodec copy {out_file}"
+    await bash(cmd)
+    if not Path(out_file).is_file():
+        return await msg.edit(get_string("audiotools_9"))
 
-    out_file = f"{file.name}.aac"
-    cmd = f"ffmpeg -i {file.name} -vn -acodec copy {out_file}"
-    o, err = await bash(cmd)
-    osremove(file.name)
     attributes = await set_attributes(out_file)
-
-    f_time = time.time()
-    try:
-        fo = await uploader(out_file, out_file, f_time, msg, f"Uploading {out_file}...")
-
-    except FileNotFoundError:
-        return await eor(msg, get_string("audiotools_9"))
-    await e.reply(
+    mmmm, _ = await e.client.fast_uploader(
+        out_file, show_progress=True, event=msg, save_cache=False
+    )
+    await reply.reply(
         get_string("audiotools_10"),
-        file=fo,
+        file=mmmm,
         thumb=ULTConfig.thumb,
         attributes=attributes,
     )
+    osremove(out_file, path)
     await msg.delete()
